Removed redundancies + fixed up chord checking (for now)
A lot of code was making barely any change to the SR so removed it to save calc time
Also, fixed up a lot of awkward coding with C# (mostly divides), and calculated the distance between chord and next note to see if it's a chord in a stream or a lone chord.
Will adjust values when roll detection code is more in-line with what is needed.