Fix LN calculations
Turns out the old code was so incorrect i think it was getting the correct figures from excluding the bad results instead of including the good results but this now has more control over LN values.
You'll see that holdaddition max value has gone from 1.0->3.5 and that's because it's unlikely for a value in there to be above 0.3 anyway.
To scale LNs alter both lnProgressionPercentage starting value and that holdaddition value.