Add support for seeing all columns utilised at current object's time
First step towards updating how chords are calculated.

Given that the only way to fully see a chord in difficulty calculations right now is to look at the chord after we reach a note on a future timestamp, this will help calculate difficulty for the current note based on the current chord instead.