Fix ICS problem
So big problem with higher keymodes is you can make chord-heavy charts with enough spacing where you can hit all of the keys and not suffer because the hitwindows of the two chords do not overlap.
The density of chords at the higher diffs means you can have quite high star rating on higher keymodes that become a mashfest and free 500+pp if not resolved.
This solves this issue for the meantime whilst a longer rework is put into place.