first crude implementation
Got down the basics of running a new hitwindow and changing the default miss window.

Things to do:
- Make it so that triggering a ghost note does not remove the note in question by cancelling the full process (requires me to map out the entire process of "hitting a note").
- Find out a way to make the base_ranges editable to be able to set HitResult.Miss to a consistent value (without setting OD).

Biggest issue here is that it's the object that processes the keypress and not the keypress that processes the object so it's kinda really annoying despite the fact it makes sense to do it this way.